fix some bugs

+++ src/main/java/com/youcode/aftas/repository/HuntingRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface HuntingRepository extends JpaRepository<Hunting, Integer> {
+public interface HuntingRepository extends JpaRepository<Hunting, Long> {
     Optional<Hunting> findByFishAndCompetitionAndMember(Fish fish, Competition competition, Member member);
     List<Hunting> findByCompetitionAndMember(Competition competition, Member member);
